Added Atm_em_counter & groups()

+++ Atm_oxo_field.cpp
@@ -11,7 +11,7 @@
  */
 
 
-Atm_oxo_field& Atm_oxo_field::begin( Atm_led_scheduler& led, uint8_t* led_map, uint8_t oxo_profile /* = PROFILE_LED */, uint8_t led_profile /* = PROFILE_LED */ ) { // Expects 9 * 3 + 2 leds
+Atm_oxo_field& Atm_oxo_field::begin( Atm_led_scheduler& led, int16_t* led_map, uint8_t oxo_profile /* = PROFILE_LED */, uint8_t led_profile /* = PROFILE_LED */ ) { // Expects 9 * 3 + 2 leds
   // clang-format off
   const static state_t state_table[] PROGMEM = { 
     /*             ON_ENTER    ON_LOOP  ON_EXIT  EVT_MATCH  EVT_1X  EVT_1O  EVT_2X  EVT_2O  EVT_3X  EVT_3O  EVT_4  EVT_5  EVT_6  EVT_7  EVT_8  EVT_9  EVT_INIT  EVT_TOGGLE  ELSE */
